fix crprod issue

+++ src/crprod.cpp
@@ -242,7 +242,7 @@ void CRprod::print_tree() const
     std::cout << ")";
 }
 
-std::string CRprod::genCode(size_t parent,size_t order,long long place,std::string indent) const
+std::string CRprod::genCode(size_t parent,long long order,long long place,std::string indent) const
 {
     std::string res;
